# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Event System Backend API
  version: 1.0.0
  description: |
    This is the backend API for Event Booking - a comprehensive event management and booking platform.
    It provides endpoints for managing events, venues, bookings, waitlists, seats, and analytics.

    ## Features
    - Event management and discovery
    - Venue and seating management
    - Real-time seat booking with hold mechanism
    - Waitlist management for sold-out events
    - Booking and cancellation handling
    - Comprehensive analytics
    - Role-based access control (USER/ADMIN)

  contact:
    name: Developer(Mit Shah)
    email: mitshah2406.work@gmail.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://evently-api.mitshah.dev.com/api/v1
    description: Production server
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: apiKey
      in: header
      name: Authorization
      description: "Enter: Bearer {your-token}"
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Invalid request"
        data:
          type: object
          nullable: true
        error:
          type: string
          example: "Validation failed"
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object

    # Auth Schemas
...
    CreateEventRequest:
      type: object
      required:
        - title
        - description
        - venue_template_id
        - start_time
        - end_time
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          example: "Summer Music Festival"
        description:
          type: string
          example: "An amazing outdoor music festival"
        venue_template_id:
          $ref: "#/components/schemas/UUID"
        start_time:
          $ref: "#/components/schemas/Timestamp"
        end_time:
          $ref: "#/components/schemas/Timestamp"
        tag_ids:
          type: array
          items:
            $ref: "#/components/schemas/UUID"
          example: []

    # Venue Schemas
...
paths:
...
  /admin/events:
    post:
      tags:
        - Admin Events
      summary: Create event (Admin)
      description: Create a new event (Admin only)
      security:
        - Bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Event"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Admin Events
      summary: Get all events (Admin)
      description: Get all events including drafts (Admin only)
      security:
        - Bearer: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Event"
  /admin/events/{eventId}:
    get:
      tags:
        - Admin Events
      summary: Get event by ID (Admin)
      description: Get event details including admin-only fields
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Event"
    put:
      tags:
        - Admin Events
      summary: Update event (Admin)
      description: Update an existing event (Admin only)
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          description: Event updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Admin Events
      summary: Delete event (Admin)
      description: Delete an event (Admin only)
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Event deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking by ID
      description: Get details of a specific booking
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Booking retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/bookings:
    get:
      tags:
        - Bookings
      summary: Get user bookings
      description: Get all bookings for the authenticated user
      security:
        - Bearer: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: status
          schema:
            type: string
            enum: ["CONFIRMED", "CANCELLED", "REFUNDED"]
      responses:
        "200":
          description: User bookings retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Booking"

  # Waitlist Endpoints
...
  /admin/events/{eventId}/cancellation-policy:
    post:
      tags:
        - Admin Cancellation
      summary: Create cancellation policy (Admin)
      description: Create a cancellation policy for an event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cancellation_deadline_hours:
                  type: integer
                  description: Hours before event when cancellation is no longer allowed
                refund_percentage:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 100
                  description: Percentage of refund offered
      responses:
        "201":
          description: Cancellation policy created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

    get:
      tags:
        - Admin Cancellation
      summary: Get cancellation policy (Admin)
      description: Get the cancellation policy for an event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Cancellation policy retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          event_id:
                            $ref: "#/components/schemas/UUID"
                          cancellation_deadline_hours:
                            type: integer
                          refund_percentage:
                            type: number

    put:
      tags:
        - Admin Cancellation
      summary: Update cancellation policy (Admin)
      description: Update the cancellation policy for an event
      security:
        - Bearer: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cancellation_deadline_hours:
                  type: integer
                refund_percentage:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 100
      responses:
        "200":
          description: Cancellation policy updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bookings/{id}/request-cancel:
    post:
      tags:
        - Cancellation
      summary: Request booking cancellation
      description: Request cancellation for a booking
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for cancellation request
      responses:
        "200":
          description: Cancellation request submitted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          cancellation_id:
                            $ref: "#/components/schemas/UUID"
                          status:
                            type: string
                          refund_amount:
                            type: number
  /cancellations/{id}:
    get:
      tags:
        - Cancellation
      summary: Get cancellation details
      description: Get details of a specific cancellation request
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Cancellation details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            $ref: "#/components/schemas/UUID"
                          booking_id:
                            $ref: "#/components/schemas/UUID"
                          status:
                            type: string
                          reason:
                            type: string
                          refund_amount:
                            type: number
                          processed_at:
                            $ref: "#/components/schemas/Timestamp"
  /users/cancellations:
    get:
      tags:
        - Cancellation
      summary: Get user cancellations
      description: Get all cancellation requests for the authenticated user
      security:
        - Bearer: []
      responses:
        "200":
          description: User cancellations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              $ref: "#/components/schemas/UUID"
                            booking_id:
                              $ref: "#/components/schemas/UUID"
                            status:
                              type: string
                            refund_amount:
                              type: number
                            created_at:
                              $ref: "#/components/schemas/Timestamp"

  # Tag Endpoints
...